example app reorg
