+++ package.json
@@ -11,7 +11,7 @@
     "test": "turbo test",
     "test-watch": "turbo test-watch",
     "format": "prettier --write \"**/*.{ts,tsx,md}\"",
-    "release": "pnpm install && pnpm build:packages && pnpm publish -r --access public --no-git-checks"
+    "release": "pnpm install && pnpm build:packages && pnpm publish -r --access public --no-git-checks --tag beta"
   },
   "devDependencies": {
     "@humeai/eslint-config": "workspace:*",
